% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{build_query}
\alias{build_query}
\title{build_query}
\usage{
build_query(...)
}
\arguments{
\item{...}{}
}
\value{
character(1L)
}
\description{
Internal function to construct urls for the api calls
}
