% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_last_update_time}
\alias{fetch_last_update_time}
\title{fetch_last_update_time}
\usage{
fetch_last_update_time(league, season, group_id)
}
\arguments{
\item{league}{[\code{character(1L)}]\cr
Shortcut specifying the desired league (e.g. bl1 for the German Bundesliga).
See \url{https://www.openligadb.de/Datenhaushalt/}.}

\item{season}{[\code{integer(1L)}]\cr
E.g. 2016}

\item{group_id}{[\code{integer(1L)}]\cr
E.g. 8 for the 8th matchday}
}
\value{
an object of class POSIXct
}
\description{
Returns date and time when data was updated to avoid unnecessary api calls
}
