% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_match}
\alias{fetch_match}
\alias{fetch_teams}
\alias{fetch_goal_getters}
\alias{fetch_current_table}
\alias{fetch_available_groups}
\alias{fetch_current_group}
\title{fetch_match}
\usage{
fetch_match(league, season, group_id, match_id)

fetch_teams(league, season)

fetch_goal_getters(league, season)

fetch_current_table(league, season)

fetch_available_groups(league, season)

fetch_current_group(league)
}
\arguments{
\item{league}{[\code{character(1L)}]\cr
Shortcut specifying the desired league (e.g. bl1 for the German Bundesliga).
See \url{https://www.openligadb.de/Datenhaushalt/}.}

\item{season}{[\code{integer(1L)}]\cr
E.g. 2016}

\item{group_id}{[\code{integer(1L)}]\cr
E.g. 8 for the 8th matchday}

\item{match_id}{[\code{integer(1L)}]\cr
Optional; If not specified returns all matches}
}
\value{
A nested list of data frames
}
\description{
MISSING MISSING
}
\section{Functions}{
\itemize{
\item \code{fetch_teams}: Get team data, returns a data.frame

\item \code{fetch_goal_getters}: Get goal scorer data, returns a data.frame

\item \code{fetch_current_table}: Get table, returns a data.frame

\item \code{fetch_available_groups}: Show available groups

\item \code{fetch_current_group}: Get most recent group
}}

