% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{parse_match}
\alias{parse_match}
\title{parse_match}
\usage{
parse_match(x)
}
\arguments{
\item{x}{[\code{list(2L)}]\cr}
}
\value{
list of two data.frames (matches & goals)
}
\description{
MISSING
}
